node add: factor out checks for joining
Factor out the code which checks if the node can join another
cluster. It will be used by the new API endpoint to join a cluster
but stays also in the CLIHandler as we keep the old legacy SSH method
for a bit.

This is not a completely 1:1 move, I changed:
* &$error(...) to $error->(...)
* removing a few empty lines, where code was so spread out that those
  lines resulted in the opposite of what they intended, i.e., less
  readability

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>