api/cluster: add join endpoint
Add an enpoint to the API which allows to join an existing PVE
cluster by only using the API instead of CLI tools (pvecm).

Use a worker as this operation may need longer than 30 seconds.
With the worker we also get a task log entry/window for an UI for
free, allowing to give better feedback.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>