lock locally on create and add
I fwe are not part of a cluster we do not need to worry about other
members messing with the config. But there may be local contenders,
e.g., two automation script instances started in parallel by mistake
or two admin (sessions) which start a create or join clsuter request
at the same time.
Reuse the local flock for this purpose.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>